2053 enable re adding of deleted workers ee v

+++ app/Http/Controllers/Service/Admin/CentreUsersController.php
@@ -9,7 +9,6 @@ use App\Http\Requests\AdminIndexCentreUsersRequest;
 use App\Http\Requests\AdminNewCentreUserRequest;
 use App\Http\Requests\AdminUpdateCentreUserRequest;
 use App\Sponsor;
-use Carbon\Carbon;
 use DB;
 use Debugbar;
 use Exception;
@@ -24,6 +23,7 @@ use League\Csv\CannotInsertRecord;
 use Log;
 use Ramsey\Uuid\Uuid;
 use Throwable;
+use function PHPUnit\Framework\isNan;
 
 class CentreUsersController extends Controller
 {
@@ -42,12 +42,12 @@ class CentreUsersController extends Controller
             'name' => 'name',
             'homeCentreArea' => 'homeCentre.sponsor.name',
             'homeCentre' => 'homeCentre.name',
-            default => function ($item) {
+            default => static function ($item) {
                 $homeCentre = $item->homeCentre;
-                return $homeCentre->sponsor->name . '#' . $homeCentre->name . '#' . $item->name;
+                return $homeCentre?->sponsor?->name . '#' . $homeCentre?->name . '#' . $item->name;
             },
         };
-        $workers = CentreUser::get()->sortBy($sorter, SORT_REGULAR, ($direction === 'desc'));
+        $workers = CentreUser::withTrashed()->get()->sortBy($sorter, SORT_REGULAR, ($direction === 'desc'));
 
         return view('service.centreusers.index', compact('workers'));
     }
@@ -70,10 +70,10 @@ class CentreUsersController extends Controller
     public function edit($id): View|Factory|Application
     {
         // Find the worker or throw a 500
-        $worker = CentreUser::findOrFail($id);
+        $worker = CentreUser::withTrashed()->findOrFail($id);
 
         // Get the homeCentreId
-        $homeCentreId = $worker->homeCentre->id;
+        $homeCentreId = $worker->homeCentre?->id;
 
         // Work out current worker-centre selections
         $workerCentres = [
@@ -253,16 +253,37 @@ class CentreUsersController extends Controller
 
     /**
      * Handle deleting a centre user
-	 * later if needed
      * @param int $id
      * @return RedirectResponse
      */
     public function delete(int $id): RedirectResponse
     {
-        $centreUser = CentreUser::findOrFail($id);
+        // must be disabled to delete
+        $centreUser = CentreUser::onlyTrashed()->findOrFail($id);
+        $name = $centreUser->name;
+        // remove any connections
         $centreUser->centre->centreUsers()->detach($id);
-        $centreUser->delete();
+        // boot them
+        $centreUser->forceDelete();
         return redirect()->route('admin.centreusers.index')
-            ->with('message', 'Worker ' . $centreUser->name . ' deleted');
+            ->with('message', 'Worker ' . $name . ' deleted');
+    }
+
+    /**
+     * Handle toggling of a centre user
+     * @param int $id
+     * @return RedirectResponse
+     */
+    public function toggle(int $id): RedirectResponse
+    {
+        $centreUser = CentreUser::withTrashed()->findOrFail($id);
+        $name = $centreUser->name;
+        if ($centreUser->deleted_at === null) {
+            $centreUser->delete();
+        } else {
+            $centreUser->restore();
+        }
+        return redirect()->route('admin.centreusers.edit', ['id' => $centreUser->id])
+            ->with('message', 'Worker ' . $name . ' updated');
     }
 }
